requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
colorama==0.4.6
enum34==1.1.10
execnet==2.1.1
Faker==24.14.0
future==1.0.0
h11==0.14.0
idna==3.7
iniconfig==2.0.0
lxml==5.2.1
namedlist==1.8
outcome==1.3.0.post0
packaging==24.0
pluggy==1.5.0
pycparser==2.22
PySocks==1.7.1
pytest==8.1.1
pytest-xdist==3.5.0
python-dateutil==2.9.0.post0
selene==2.0.0rc9
selenium==4.20.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
wsproto==1.2.0

